+++ js/renderLines.js
@@ -1,6 +1,6 @@
 // renderLines.js
-// 表示エリアの描画とドラッグ＆ドロップ処理を担当
-import { getLines, moveLine } from './linesData.js';
+// 表示エリアの描画のみを担当
+import { getLines } from './linesData.js';
 
 export function renderLines(displayArea) {
     const linesData = getLines();
@@ -14,34 +14,4 @@ export function renderLines(displayArea) {
         }
         return `<div class='palette-line' draggable='true' data-idx='${idx}'>${content}</div>`;
     }).join('');
-
-    let dragSrcIdx = null;
-    displayArea.querySelectorAll('.palette-line').forEach(lineEl => {
-        lineEl.addEventListener('dragstart', (e) => {
-            dragSrcIdx = parseInt(lineEl.getAttribute('data-idx'));
-            e.dataTransfer.effectAllowed = 'move';
-            lineEl.classList.add('dragging');
-        });
-        lineEl.addEventListener('dragend', (e) => {
-            lineEl.classList.remove('dragging');
-        });
-        lineEl.addEventListener('dragover', (e) => {
-            e.preventDefault();
-            e.dataTransfer.dropEffect = 'move';
-            lineEl.classList.add('drag-over');
-        });
-        lineEl.addEventListener('dragleave', (e) => {
-            lineEl.classList.remove('drag-over');
-        });
-        lineEl.addEventListener('drop', (e) => {
-            e.preventDefault();
-            lineEl.classList.remove('drag-over');
-            const targetIdx = parseInt(lineEl.getAttribute('data-idx'));
-            if (dragSrcIdx !== null && dragSrcIdx !== targetIdx) {
-                moveLine(dragSrcIdx, targetIdx);
-                renderLines(displayArea);
-            }
-            dragSrcIdx = null;
-        });
-    });
 }
